Introduce websocket module to python
