+++ python/foxglove-sdk/python/foxglove/websocket.py
@@ -1,0 +1,31 @@
+from ._foxglove_py.websocket import (
+    Capability,
+    ChannelView,
+    Client,
+    ConnectionGraph,
+    MessageSchema,
+    Parameter,
+    ParameterType,
+    ParameterValue,
+    Service,
+    ServiceRequest,
+    ServiceSchema,
+    StatusLevel,
+    WebSocketServer,
+)
+
+__all__ = [
+    "Capability",
+    "ChannelView",
+    "Client",
+    "ConnectionGraph",
+    "MessageSchema",
+    "Parameter",
+    "ParameterType",
+    "ParameterValue",
+    "Service",
+    "ServiceRequest",
+    "ServiceSchema",
+    "StatusLevel",
+    "WebSocketServer",
+]
